Add toggle for visualization functionality
This uses MOOSE's Controls system so the user only needs to toggle
the value of 'visualization' to disable auxkernels and exodus output.
This is not as efficient as using HIT's (in)active flags, because
the objects are still created just not executed. I do not know of
any way of achieving the same one-parameter toggle functionality
within HIT alone.